Neat.  We're generating lists of content now.
We have a tree walker combined with some light nifty functional
composition that gives us a list of property lists.  Each property
list corresponds to one file of content discovered in a tree.  This
metadata is a combination of configured defaults and values specified
at the top of each content file, all of which is then shoved through
functions that clean up text, parse dates, and the like.

Get the show going:
    (DISCOVER-CONTENT "testsrc")

Sort the metadata as you probably would for a blog.
    (SORT * #'< :key #'(lambda (x) (getf x 'date)))

I'm shocked how easy this is, so far.