Added FMT to meta data.
We examine the pathname and determine a FMT for the content, when we
can.  We'll catch unsupported things and other errors later during
processing; discover-content really shouldn't bark much at all.

I think I should call this good, and focus on the config file and the
command line now.  Write something that displays the metadata for each
bit of content, hook that into main along with discover-content, and
go from there.