Started to explore executable... whoa, Nelly.
Looked at

- buildapp (which I sorta remember)
- trivial-build (stale if not abandoned after 2015)
- cl-launch (rings a bell, but looking it over, I can see why I never
  poked at it much)
- roswell (now with 50% more crazy juice!)
- clasp (!)

I think Roswell is the right path forward, it'll do what we need and
it'll support Windows and Unix-derived operating systems across quite
a number of Lisp environments.  But, wow, that's a LOT of
infrastructure for such a small job.

I'm still tempted to play with scripts and <os>.lisp, but I think that
is its own special flavor of madness.  I shouldn't be wasting time on
application delivery.